+++ packages/app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/app
 
+## 0.5.15
+
+### Patch Changes
+
+- @tinyhttp/res@0.5.5
+
 ## 0.5.14
 
 ### Patch Changes