+++ packages/ip-filter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/ip-filter
 
+## 0.5.15
+
+### Patch Changes
+
+- @tinyhttp/app@0.5.15
+
 ## 0.5.14
 
 ### Patch Changes